Translate infer_bin
