+++ tests/tests.rs
@@ -8,192 +8,10 @@ use rls::lsp_data::request::Request as _;
 
 use self::support::harness::compare_json;
 use self::support::project_builder::{project, ProjectBuilder};
-use self::support::{basic_bin_manifest, fixtures_dir, rls_timeout, RlsStdout};
+use self::support::{fixtures_dir, rls_timeout, RlsStdout};
 
 #[allow(dead_code)]
 mod support;
-
-#[test]
-fn cmd_test_infer_bin() {
-    let p = project("simple_workspace")
-        .file("Cargo.toml", &basic_bin_manifest("foo"))
-        .file(
-            "src/main.rs",
-            r#"
-                struct UnusedBin;
-                fn main() {
-                    println!("Hello world!");
-                }
-            "#,
-        )
-        .build();
-
-    let root_path = p.root();
-    let mut rls = p.spawn_rls();
-
-    rls.request(
-        0,
-        "initialize",
-        Some(json!({
-            "rootPath": root_path,
-            "capabilities": {}
-        })),
-    )
-    .unwrap();
-
-    let json: Vec<_> = rls
-        .wait_until_done_indexing(rls_timeout())
-        .to_json_messages()
-        .filter(|json| json["method"] != "window/progress")
-        .collect();
-
-    assert!(json.len() > 1);
-
-    assert!(json[0]["result"]["capabilities"].is_object());
-
-    assert_eq!(json[1]["method"], "textDocument/publishDiagnostics");
-    assert_eq!(json[1]["params"]["diagnostics"][0]["code"], "dead_code");
-
-    rls.shutdown(rls_timeout());
-}
-
-/// Test includes window/progress regression testing
-#[test]
-fn cmd_test_simple_workspace() {
-    let p = project("simple_workspace")
-        .file(
-            "Cargo.toml",
-            r#"
-                [workspace]
-                members = [
-                "member_lib",
-                "member_bin",
-                ]
-            "#,
-        )
-        .file(
-            "Cargo.lock",
-            r#"
-                [root]
-                name = "member_lib"
-                version = "0.1.0"
-
-                [[package]]
-                name = "member_bin"
-                version = "0.1.0"
-                dependencies = [
-                "member_lib 0.1.0",
-                ]
-            "#,
-        )
-        .file(
-            "member_bin/Cargo.toml",
-            r#"
-                [package]
-                name = "member_bin"
-                version = "0.1.0"
-                authors = ["Igor Matuszewski <Xanewok@gmail.com>"]
-
-                [dependencies]
-                member_lib = { path = "../member_lib" }
-            "#,
-        )
-        .file(
-            "member_bin/src/main.rs",
-            r#"
-                extern crate member_lib;
-
-                fn main() {
-                    let a = member_lib::MemberLibStruct;
-                }
-            "#,
-        )
-        .file(
-            "member_lib/Cargo.toml",
-            r#"
-                [package]
-                name = "member_lib"
-                version = "0.1.0"
-                authors = ["Igor Matuszewski <Xanewok@gmail.com>"]
-
-                [dependencies]
-            "#,
-        )
-        .file(
-            "member_lib/src/lib.rs",
-            r#"
-                pub struct MemberLibStruct;
-
-                struct Unused;
-
-                #[cfg(test)]
-                mod tests {
-                    #[test]
-                    fn it_works() {
-                    }
-                }
-            "#,
-        )
-        .build();
-
-    let root_path = p.root();
-    let mut rls = p.spawn_rls();
-
-    rls.request(
-        0,
-        "initialize",
-        Some(json!({
-            "rootPath": root_path,
-            "capabilities": {}
-        })),
-    )
-    .unwrap();
-
-    let json: Vec<_> = rls
-        .wait_until_done_indexing(rls_timeout())
-        .to_json_messages()
-        .collect();
-    assert!(json.len() >= 11);
-
-    assert!(json[0]["result"]["capabilities"].is_object());
-
-    assert_eq!(json[1]["method"], "window/progress");
-    assert_eq!(json[1]["params"]["title"], "Building");
-    assert_eq!(json[1]["params"].get("message"), None);
-
-    // order of member_lib/member_bin is undefined
-    for json in &json[2..6] {
-        assert_eq!(json["method"], "window/progress");
-        assert_eq!(json["params"]["title"], "Building");
-        assert!(json["params"]["message"]
-            .as_str()
-            .unwrap()
-            .starts_with("member_"));
-    }
-
-    assert_eq!(json[6]["method"], "window/progress");
-    assert_eq!(json[6]["params"]["done"], true);
-    assert_eq!(json[6]["params"]["title"], "Building");
-
-    assert_eq!(json[7]["method"], "window/progress");
-    assert_eq!(json[7]["params"]["title"], "Indexing");
-
-    assert_eq!(json[8]["method"], "textDocument/publishDiagnostics");
-
-    assert_eq!(json[9]["method"], "textDocument/publishDiagnostics");
-
-    assert_eq!(json[10]["method"], "window/progress");
-    assert_eq!(json[10]["params"]["done"], true);
-    assert_eq!(json[10]["params"]["title"], "Indexing");
-
-    let json = rls
-        .shutdown(rls_timeout())
-        .to_json_messages()
-        .nth(11)
-        .expect("No shutdown response received");
-
-    assert_eq!(json["id"], 99999);
-}
 
 #[test]
 fn cmd_changing_workspace_lib_retains_bin_diagnostics() {
